Separate `aiohttp.http_websocket` into multiple files (#9542)
(cherry picked from commit dd9a1fdf568de33bc444779ca5a273713223666c)
