+++ aiohttp/_websocket/__init__.py
@@ -1,0 +1,1 @@
+"""WebSocket protocol versions 13 and 8."""
