+++ tests/test_websocket_writer.py
@@ -141,7 +141,7 @@ async def test_concurrent_messages(
       and in the executor
     """
     with mock.patch(
-        "aiohttp.http_websocket.WEBSOCKET_MAX_SYNC_CHUNK_SIZE", max_sync_chunk_size
+        "aiohttp._websocket.writer.WEBSOCKET_MAX_SYNC_CHUNK_SIZE", max_sync_chunk_size
     ):
         writer = WebSocketWriter(protocol, transport, compress=15)
         queue: DataQueue[WSMessage] = DataQueue(asyncio.get_running_loop())
